% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_pow_curve.R
\name{cluster_random_sample_size}
\alias{cluster_random_sample_size}
\title{Cluster Random Sample Size}
\usage{
cluster_random_sample_size(alpha, power, coef_var = 0.25, A = 2, pi, pc, CS)
}
\arguments{
\item{alpha}{significance level}

\item{power}{the desired power}

\item{coef_var}{between-cluster variation, use 0.25 if unknown as this is the estimated maximum}

\item{A}{Adjustment for loss of degrees of freedom (1 for unpaired trials, 2 in paired trials)}

\item{pi}{proportion in the intervention group}

\item{pc}{proportion in the control group}

\item{CS}{an estimate of the cluster size}
}
\value{
number of clusters required in each treatment group in an unmatched design, and total pairs in a paired design.
}
\description{
Gives the number of clusters (pairs of clusters for paired designs) required for cluster randomized trials
}
\examples{

# The matched trial protocol required 80\% power of detecting a 50\% reduction of annual HIV incidence from an assumed 1\% in the control arm (π0 = 0.02 over 2 years) to 0.5\% in the intervention arm (π1 = 0.01). Thus, with m = 1000 adults followed up in each cluster, the number of clusters required per treatment arm in a matched design was gives 6.8 = 7 cluster pairs
cluster_random_sample_size(alpha=0.05, power=0.8, coef_var=0.25, A =2, pi=0.01, pc=0.02, CS=1000)
}
\references{
Hayes, R. J., & Moulton, L. H. (2017). Cluster randomised trials. Chapman and Hall/CRC.
}
