% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_proportions_cross.R
\name{compare_proportions_cross}
\alias{compare_proportions_cross}
\title{compare_proportions_cross}
\usage{
compare_proportions_cross(alpha = 0.05, power = 0.8, p_diff, var = 0.5)
}
\arguments{
\item{alpha}{significance level, default 5\%}

\item{power}{desired power, default 80\%}

\item{p_diff}{difference between the proportions of incidence}

\item{var}{variance of the d}
}
\value{

}
\description{
Sample size calculation to compare proportions in a cross-over design
}
\examples{
# Difference between the two proportions are predicted to be 10\%
compare_proportions_cross(alpha = 0.05, power = 0.8, p_diff = 0.1, var = 0.5)
}
\references{
Chow, S. C., Shao, J., Wang, H., & Lokhnygina, Y. (2017). Sample size calculations in clinical research. chapman and hall/CRC.
}
