% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_pow_curve.R
\name{effective_sample_size}
\alias{effective_sample_size}
\title{Effective Sample Size}
\usage{
effective_sample_size(m, k, ICC)
}
\arguments{
\item{m}{number of participants in a cluster}

\item{k}{number of clusters}

\item{ICC}{intracluster correlation coefficient}
}
\value{
effective sample size (actual sample size (mk) divided by the design effect (1 + ICC * (m - 1)))
}
\description{
Gives the 'real' sample size in a clustered trial, accounting for the net loss of data due to the similarities among participants in the same cluster.
}
\examples{

# In a cluster randomized controlled trial, you have 8 clusters with 30 participants per cluster i.e., total sample size of 240 and an ICC of 0.2. A simple method to calculate the power associated with this sample size is to first calculate the effective sample size i.e., the sample size in a simple random sample that would give the same level of precision, and then do a power calculation using the ESS. 

effective_sample_size(m = 30, k =8, ICC = 0.2)

}
\references{
Killip S, Mahfoud Z, Pearce K. (2004). What is an intracluster correlation coefficient? Crucial concepts for primary care researchers. Ann Fam Med.
}
