% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findsamplesize-package.R
\docType{package}
\name{findsamplesize-package}
\alias{findsamplesize}
\alias{findsamplesize-package}
\title{findsamplesize: Sample Size Calculations}
\description{
Gives the number of participants necessary for the RDT performance evaluation studies.
}
\author{
\strong{Maintainer}: Berra Erkosar \email{berra.erkosar@finddx.org}

Other contributors:
\itemize{
  \item FIND [copyright holder]
}

}
\keyword{internal}
