% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_pow_curve.R
\name{n_pop_prop}
\alias{n_pop_prop}
\title{Sample Size Population Proportion}
\usage{
n_pop_prop(alpha, p, margin, N = NULL)
}
\arguments{
\item{alpha}{numeric, significance level}

\item{p}{numeric, estimated proportion}

\item{margin}{numeric, error margin}

\item{N}{numeric, population size}
}
\value{
sample size for the estimation of the proportion
}
\description{
Sample Size Population Proportion
}
\examples{

 n_pop_prop(alpha = 0.05, p =  0.5, margin = 0.1, N = 1000) 

# sample size for a population of 1000 people to estimate that at least half of the population has the desired characteristic with an error margin of 10\%

 n_pop_prop(alpha = 0.05, p =  0.5, margin = 0.1) 
 # same calculation in an infinite population
}
\references{
Daniel  WW  (1999).  Biostatistics:  A  Foundation for   Analysis   in   the   Health   Sciences.   7th edition. New York: John Wiley & Sons.
}
