% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_calculation.R
\name{n_tot_calculation}
\alias{n_tot_calculation}
\title{Total N Calculation}
\usage{
n_tot_calculation(prevalence, sample_size, prevalence_power)
}
\arguments{
\item{prevalence}{Disease prevalence}

\item{sample_size}{Number of confirmed cases by reference}

\item{prevalence_power}{the power to detect the prevalence}
}
\value{
total number of subjects to be screened
}
\description{
Total N Calculation
}
\examples{

 n_tot_calculation(sample_size=503, prevalence = 0.1, prevalence_power = 0.9) 
# Total  number of patients to be screened to test 80\% sensitivity with 10\% prevalence and 80\% prevalence power. 
# Should give the same number as the second example of sample_size_calculation

}
\references{
Zhou XH, Obuchowski NA and McClish DK. Statistical Methods in Diagnostic Medicine. 2011;2:193-228
}
