% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_calculation.R
\name{power_calc}
\alias{power_calc}
\title{Power Calculation}
\usage{
power_calc(sample_size, margin, sen_spe, alpha)
}
\arguments{
\item{sample_size}{numeric, confirmed positives/negatives}

\item{margin}{numeric, the error margin}

\item{sen_spe}{test performance characteristice estimate}

\item{alpha}{significance level}
}
\value{

}
\description{
Power Calculation
}
\examples{

 power_calc(alpha=0.05, margin=0.05, sen_spe=0.8, sample_size=502) 
# Power of the study. Using the same parameters as sample_size_calculation. Should be ~80\%

}
\references{
Zhou XH, Obuchowski NA and McClish DK. Statistical Methods in Diagnostic Medicine. 2011;2:193-228
}
