% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_calculation.R
\name{sample_size_calculation}
\alias{sample_size_calculation}
\title{sample_size_calculation}
\usage{
sample_size_calculation(
  sen_spe,
  alpha,
  power,
  margin,
  prevalence = NULL,
  prevalence_power = NULL,
  performance_characteristic = "sen"
)
}
\arguments{
\item{sen_spe}{numeric, sensitivity/specificity}

\item{alpha}{numeric, significance level}

\item{power}{numeric, desired power}

\item{margin}{numeric, the error margin}

\item{prevalence}{numeric, prevalence of the disease. For specificity put 1-diesease prevalance}

\item{prevalence_power}{numeric, the power to detect the prevalence}

\item{performance_characteristic}{character, "sen" for sensitivity, "spe" for specificity}
}
\value{
sample size
}
\description{
sample_size_calculation
}
\examples{
 sample_size_calculation(sen_spe= 0.8, alpha=0.05, power=0.8, margin=0.05, prevalence = NULL, prevalence_power = NULL, performance_characteristic = "sen") 
# Number of true positives for a sensitivity of 80\%

 sample_size_calculation(sen_spe= 0.8, alpha=0.05, power=0.8, margin=0.05, prevalence = 0.1, prevalence_power = 0.9, performance_characteristic = "sen") 
 # Total  number of patients to be screened to test 80\% sensitivity with 10\% prevalence and 80\% prevalence power


}
\references{
Zhou XH, Obuchowski NA and McClish DK. Statistical Methods in Diagnostic Medicine. 2011;2:193-228
}
