% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_compare_sen_spe.R
\name{ss_compare_sen_spe}
\alias{ss_compare_sen_spe}
\title{Sample Size for detecting a difference between two tests}
\usage{
ss_compare_sen_spe(theta_s, theta_e, delta_m, alpha = 0.05, beta = 0.8)
}
\arguments{
\item{theta_s}{numeric value between 0 and 1. performance estimate of standard test}

\item{theta_e}{numeric value between 0 and 1. performance estimate of experimental test}

\item{delta_m}{numeric value. maximal acceptable difference to conclude non-inferiority.}

\item{alpha}{numeric value. Significance level}

\item{beta}{numeric value. beta = 1 - Power}
}
\value{
numeric values giving the min-mean-max number of samples required
}
\description{
Sample size calculation when comparing sensitivities/specifies in paired groups by McNemar, according to Beam, 1992, Table 4. Three sample sizes are given taking into account the minimum (N_min) and the maximum (N_max) bounds on the probability of disagreement, and their mean (N_mean). The original formula is for equivalence testing. The hypothesis of detecting a specified difference (delta_m) was done as in Blackwleder 1982 by introducing the specified difference to the denominator.
}
\examples{
\dontrun{
ss_compare_sen_spe(theta_s = 0.45, theta_e = 0.43, delta_m = 0.07, beta = 0.8)
ss_compare_sen_spe(theta_s = 0.45, theta_e = 0.43, delta_m = 0, beta = 0.8)
}
}
\references{
Blackwelder WC. "Proving the Null Hypothesis" in Clinical Trials. Control. Clin. Trials 1982; 3:345-353.

Beam, C. A. (1992). Strategies for improving power in diagnostic radiology research. American Journal of Roentgenology 159, 631-637.
}
