% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_non_inf_2.R
\name{ss_non_inf_2}
\alias{ss_non_inf_2}
\title{Sample Size non-inferiority method 2}
\usage{
ss_non_inf_2(theta_s, theta_e, delta_m, alpha = 0.05, beta = 0.8, p11 = NULL)
}
\arguments{
\item{theta_s}{numeric value between 0 and 1. performance estimate of standard test}

\item{theta_e}{numeric value between 0 and 1. performance estimate of experimental test}

\item{delta_m}{numeric value. maximal acceptable difference to conclude non-inferiority.}

\item{alpha}{numeric value. Significance level}

\item{beta}{numeric value. beta = 1 - Power}

\item{p11}{numeric value. Probability that standard test is positive when experimental test is positive, when estimation is possible from preliminary data}
}
\value{
numeric value giving the number of samples required
}
\description{
Gives a sample size  for a diagnostic accuracy trial with non-inferiority hypothesis, based on estimated accuracy for each test and a minimum non-acceptable difference (Eq. 6.24 from Zhou et al. 2011)  Variance is calculated using the Eq.6.15 from Zhou et al. In default version, the variance is calculated twice for the case when there is perfect correlation between the test results and the case when the correlation between the tests are equal to 0. A sample size for both scenarios and their mean is provided. In case of unknown correlation with a "modearate" expectation, mean value can be used.
}
\examples{
\dontrun{
ss_non_inf_2(theta_s = 0.45, theta_e = 0.43, delta_m = 0.07, beta = 0.8)
ss_non_inf_2(theta_s = 0.45, theta_e = 0.43, delta_m = 0.07, beta = 0.8, p11 = 0.5)
}
}
